restructed some of the components inside the rooms directory and components directory.  Got started on the Rooms.js component inside of the Rooms directory inside of the pages folder
