+++ client/src/Pages/Main/Main.css
@@ -1,7 +1,7 @@
 /* MAIN TITLE SECTION */
 
 #mainTitleSection{
-   background-image: url('../../Images/HotelExteriorTwo.jpg');
+   background-image: url('../../Images/HotelExteriorThree.jpg');
    height: 750px; 
    background-size: cover;
 }
